requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
alluka==0.3.3
annotated-types==0.7.0
anyio==4.10.0
attrs==24.3.0
cffi==2.0.0
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
cryptography==46.0.3
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
h11==0.16.0
hikari==2.0.0
hikari-lightbulb==2.3.5
hikari-miru==4.2.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.5.1
pillow==11.2.1
propcache==0.3.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.2.1
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
yarl==1.20.1
